Изменена запись плагина для pom.xml

+++ src/main/java/ru/netology/bonus/BonusService.java
@@ -1,7 +1,7 @@
 package ru.netology.bonus;
 
 public class BonusService {
-    public long calculate(long amount, boolean registered) {
+    public long calculate(int amount, boolean registered) {
         int percent = registered ? 3 : 1;
         long bonus = amount * percent / 100 / 100;
         long limit = 500;
