+++ src/test/java/ru/netology/bonus/BonusServiceTest.java
@@ -1,16 +1,14 @@
 package ru.netology.bonus;
 
 import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvFileSource;
-import org.junit.jupiter.params.provider.CsvSource;
 
 public class BonusServiceTest {
 
     @ParameterizedTest
     @CsvFileSource (resources = "/data.csv")
-    public void test(long amount, boolean registered, long expected) {
+    public void test(int amount, boolean registered, long expected) {
         BonusService service = new BonusService();
         long actual = service.calculate(amount, registered);
